Add TLK boot arguments
TLK needs to know where the kernel is located. This modifies the
existing small ARM bootloader and adds an extra variable when booting
TrustZone images so they can get an argument in r0.

hw/arm_boot.c: Change tz boot code

The tz image now returns to the bootloader and the bootloader
starts the kernel.

Change-Id: I03eb8fd7e96ead756cc9d88c0bec8b21fafb56ee